Fix: Get Vendor logic CCT
